done disktra begin MIS

+++ src/algoObjects/Node.java
@@ -4,12 +4,12 @@ import java.awt.Color;
 
 public class Node {
 	private Color color;
-	private double coordinate_x;
-	private double coordinate_y;
-	private double radius;
+	private int coordinate_x;
+	private int coordinate_y;
+	private int radius;
 	
 	public static int numberOfNodes = 0;
-	public Node(double coordinate_x, double coordinate_y, Color color, double radius) {
+	public Node(int coordinate_x, int coordinate_y, Color color, int radius) {
 		this.color = color;
 		this.coordinate_x = coordinate_x;
 		this.coordinate_y = coordinate_y;
@@ -20,15 +20,15 @@ public class Node {
 		this.color = color;
 	}
 	
-	public double getX() {
+	public int getX() {
 		return this.coordinate_x;
 	}
 	
-	public double getY() {
+	public int getY() {
 		return this.coordinate_y;
 	}
 	
-	public double getRadius() {
+	public int getRadius() {
 		return this.radius;
 	}
